Set block_size to u64

+++ src/ot/bitmatrix.rs
@@ -13,7 +13,7 @@ use std::ops::Index;
 // BitMatrix and BitVector
 use serde::{Deserialize, Serialize};
 
-pub type Block = u8;
+pub type Block = u64;
 pub const BLOCK_SIZE: usize = mem::size_of::<Block>() * 8;
 
 #[repr(transparent)]
